move styles to separate folder and file;
